Improve Unit Tests for basic types, test 32 and 64 bit integer types as well.
