+++ tests/helpers/__init__.py
@@ -3,11 +3,12 @@ Test helpers
 """
 
 from .optional_exception_tester import OptionalExceptionTester
-from .helper_functions import cartesianproduct, random_sample, create_testset_simple_range
+from .helper_functions import cartesianproduct, random_sample, create_testset_simple_range, check_tag
 
 __all__ = [
         "OptionalExceptionTester",
         "cartesianproduct",
         "create_testset_simple_range",
+        "check_tag",
         "random_sample",
         ]
